fix: improve code
