+++ apps/client/src/pages/Dashboard/Groups/GroupDetails/components/ContributionOverview/cardComponent.module.scss
@@ -27,5 +27,3 @@
     width: 40%;
   }
 }
-
-
