+++ apps/client/src/pages/Dashboard/Groups/GroupDetails/components/ContributionOverview/mockData.ts
@@ -1,0 +1,34 @@
+export const viewData = [
+  {
+    label: 'Contribution Amount',
+    value: '₦ 500,000',
+  },
+  {
+    label: 'Schedule',
+    value: 'Daily',
+  },
+  {
+    label: 'Contribution Timeline',
+    value: '5 months',
+  },
+  {
+    label: 'Estimated Collection',
+    value: '₦ 2,000,000',
+  },
+  {
+    label: 'Start Date',
+    value: 'May 1, 2022',
+  },
+  {
+    label: 'End Date',
+    value: 'October 1, 2022',
+  },
+  {
+    label: 'Available Slots',
+    value: '2 and 6',
+  },
+  {
+    label: 'Total Slots',
+    value: '6',
+  },
+]
