+++ apps/client/src/pages/Dashboard/Groups/GroupDetails/components/MembersPerformance/helpers.ts
@@ -1,0 +1,7 @@
+export const getPerformanceColor = (performance: string | number): string => {
+  if (typeof performance === 'string') {
+    return 'var(--primary-100)'
+  } else {
+    return performance >= 50 ? 'green' : 'red'
+  }
+}
