+++ apps/client/src/pages/Dashboard/Groups/GroupDetails/components/MembersPerformance/tableComponent.module.scss
