+++ apps/client/src/pages/Dashboard/Groups/GroupDetails/exploreGroup.module.scss
