+++ apps/client/src/pages/Dashboard/Groups/exploreGroups.tsx
@@ -1,4 +1,5 @@
 import ExploreGroup from '../../../components/compounds/ExploreGroups/ExploreGroups'
+
 const ExploreGroups = () => {
   return (
     <div>
