+++ apps/client/src/pages/Dashboard/Groups/index.tsx
@@ -7,8 +7,6 @@ const GroupsPage = () => {
   const params = new URLSearchParams(location.search)
   const currentPage = params.get('page')
 
-  console.log({ currentPage })
-
   if (currentPage === 'explore') {
     return <ExploreGroups />
   }
